# openapi.yaml
openapi: 3.0.3
info:
  title: Alora Backend API
  description: >
    API REST del backend de Alora, una plataforma de gestión de cuidados geriátricos.
    Gestiona autenticación JWT, perfiles de pacientes, registros de cuidado,
    recordatorios y un asistente IA basado en Gemini.
  version: 1.0.0
  contact:
    name: Jose Rodriguez
servers:
  - url: http://localhost:8080
    description: Servidor de desarrollo local
security:
  - BearerAuth: []
...
paths:
...
  /api/admin/users:
    get:
      tags: [Admin]
      summary: Listar todos los usuarios del sistema
      description: >
        Devuelve la lista de cuentas registradas sin exponer contraseñas.
        Requiere rol **ADMIN** — un usuario con rol USER recibirá 403.
      responses:
        '200':
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserSummaryDto'
              example:
                - id: 1
                  email: "admin@alora.com"
                  fullName: "Admin Principal"
                  role: "ADMIN"
                - id: 2
                  email: "cuidadora@ejemplo.com"
                  fullName: "María García"
                  role: "USER"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/admin/users/{id}:
    get:
      tags: [Admin]
      summary: Obtener un usuario por ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Usuario encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSummaryDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags: [Admin]
      summary: Eliminar un usuario
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '204':
          description: Usuario eliminado correctamente
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/admin/users/{id}/role:
    patch:
      tags: [Admin]
      summary: Cambiar el rol de un usuario
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleRequest'
            example:
              role: "ADMIN"
      responses:
        '200':
          description: Rol actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSummaryDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ── PROFILES ──────────────────────────────────────────────
  /api/profiles:
    get:
      tags: [Profiles]
      summary: Listar mis perfiles de pacientes
      description: Devuelve únicamente los perfiles creados por el usuario autenticado.
      responses:
        '200':
          description: Lista de perfiles del cuidador
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PrivateProfileDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags: [Profiles]
      summary: Crear un nuevo perfil de paciente
      description: >
        Crea la ficha de un paciente. El sistema genera automáticamente un
        `qrToken` único. El `pinCode` se usa para proteger el acceso público.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileRequest'
            example:
              fullName: "Juan Pérez López"
              email: "juan@ejemplo.com"
              birthDate: "1945-03-15"
              gender: "Masculino"
              medicalConditions: "Diabetes tipo 2, Hipertensión arterial"
              allergies: "Penicilina"
              medications: "Metformina 850mg, Enalapril 10mg"
              approxAddress: "Calle Mayor 10, 2º B"
              city: "Madrid"
              emergencyContactName: "Ana Pérez"
              emergencyContactPhone: "600123456"
              emergencyContactEmail: "ana@ejemplo.com"
              relationship: "Hija"
              pinCode: "1234"
      responses:
        '200':
          description: Perfil creado con `qrToken` generado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivateProfileDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/profiles/{profileId}/logs:
    get:
      tags: [Care Logs]
      summary: Obtener registros de cuidado de un perfil
      description: >
        Devuelve el historial de cuidados ordenado del más reciente al más antiguo.
        Solo el dueño del perfil puede acceder a sus registros.
      parameters:
        - $ref: '#/components/parameters/ProfileIdPath'
      responses:
        '200':
          description: Lista de registros ordenada por fecha descendente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CareLogDto'
              example:
                - id: 10
                  profileId: 5
                  logType: "MEDICACION"
                  note: "Tomó ibuprofeno 400mg con el desayuno"
                  createdAt: "2024-01-15T09:00:00"
                - id: 9
                  profileId: 5
                  logType: "COMIDA"
                  note: "Desayunó tostadas con aceite"
                  createdAt: "2024-01-15T08:30:00"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags: [Care Logs]
      summary: Crear un nuevo registro de cuidado
      description: >
        Añade un evento al historial del paciente.
        El campo `id` en el body es ignorado — se asigna siempre uno nuevo.
      parameters:
        - $ref: '#/components/parameters/ProfileIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CareLogRequest'
            example:
              logType: "MEDICACION"
              note: "Se administró Metformina 850mg con el desayuno"
      responses:
        '200':
          description: Registro de cuidado creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CareLogDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/profiles/{profileId}/logs/{logId}:
    get:
      tags: [Care Logs]
      summary: Obtener un registro de cuidado por ID
      parameters:
        - $ref: '#/components/parameters/ProfileIdPath'
        - name: logId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Registro encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CareLogDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags: [Care Logs]
      summary: Actualizar un registro de cuidado
      description: Solo se actualizan los campos `logType` y `note` — el resto es inmutable.
      parameters:
        - $ref: '#/components/parameters/ProfileIdPath'
        - name: logId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CareLogRequest'
            example:
              logType: "MEDICACION"
              note: "Tomó ibuprofeno 600mg (dosis corregida)"
      responses:
        '200':
          description: Registro actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CareLogDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags: [Care Logs]
      summary: Eliminar un registro de cuidado
      parameters:
        - $ref: '#/components/parameters/ProfileIdPath'
        - name: logId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '204':
          description: Registro eliminado correctamente
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ── REMINDERS ─────────────────────────────────────────────
  /api/profiles/{profileId}/reminders:
    get:
      tags: [Reminders]
      summary: Obtener recordatorios de un perfil
      description: Devuelve los recordatorios ordenados por hora ascendente (primero el más temprano).
      parameters:
        - $ref: '#/components/parameters/ProfileIdPath'
      responses:
        '200':
          description: Lista de recordatorios ordenada por hora
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReminderDto'
              example:
                - id: 3
                  title: "Pastilla tensión"
                  time: "08:30:00"
                  isActive: true
                  daysOfWeek: "LUNES,MIERCOLES,VIERNES"
                - id: 4
                  title: "Insulina"
                  time: "20:00:00"
                  isActive: true
                  daysOfWeek: "TODOS"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags: [Reminders]
      summary: Crear un recordatorio
      description: >
        Si `daysOfWeek` no se especifica, se asigna "TODOS" por defecto.
        Los días se pasan como una cadena separada por comas (ej. "LUNES,MARTES").
      parameters:
        - $ref: '#/components/parameters/ProfileIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReminderRequest'
            example:
              title: "Pastilla tensión"
              time: "08:30:00"
              daysOfWeek: "LUNES,MIERCOLES,VIERNES"
      responses:
        '200':
          description: Recordatorio creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/profiles/{profileId}/reminders/{reminderId}:
    get:
      tags: [Reminders]
      summary: Obtener un recordatorio por ID
      parameters:
        - $ref: '#/components/parameters/ProfileIdPath'
        - name: reminderId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Recordatorio encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags: [Reminders]
      summary: Actualizar un recordatorio
      parameters:
        - $ref: '#/components/parameters/ProfileIdPath'
        - name: reminderId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReminderRequest'
            example:
              title: "Pastilla tensión (nueva dosis)"
              time: "09:00:00"
              daysOfWeek: "TODOS"
      responses:
        '200':
          description: Recordatorio actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags: [Reminders]
      summary: Eliminar un recordatorio
      parameters:
        - $ref: '#/components/parameters/ProfileIdPath'
        - name: reminderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID del recordatorio a eliminar
          example: 3
      responses:
        '204':
          description: Recordatorio eliminado correctamente
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ── PUBLIC PROFILE ────────────────────────────────────────
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Token JWT obtenido en `/auth/login` o `/auth/register`.
        Enviar como: `Authorization: Bearer <token>`

  # ── PARAMETERS ────────────────────────────────────────────
  parameters:
...
    ProfileIdPath:
      name: profileId
      in: path
      required: true
      description: ID del perfil de paciente
      schema:
        type: integer
        format: int64
        example: 5
...
  responses:
    Unauthorized:
      description: Token JWT ausente, expirado o inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: "2024-01-15T10:30:00Z"
            status: 401
            error: "Unauthorized"
            message: "Full authentication is required to access this resource"
            path: "/api/profiles"
    Forbidden:
      description: El usuario autenticado no tiene permisos suficientes (rol incorrecto o recurso ajeno)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: "2024-01-15T10:30:00Z"
            status: 403
            error: "Forbidden"
            message: "Access Denied"
            path: "/api/admin/users"
    NotFound:
      description: El recurso solicitado no existe
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: "2024-01-15T10:30:00Z"
            status: 404
            error: "Not Found"
            message: "Perfil no encontrado"
            path: "/api/profiles/99"
...
    ValidationError:
      description: Uno o más campos no superan las restricciones de validación
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
          example:
            timestamp: "2024-01-15T10:30:00Z"
            status: 400
            error: "Validation Error"
            message: "Los datos enviados no son válidos"
            path: "/auth/register"
            validationErrors:
              email: "Debe ser un email valido"
              password: "La contraseña debe tener al menos 6 caracteres"
    InternalServerError:
      description: Error interno inesperado del servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: "2024-01-15T10:30:00Z"
            status: 500
            error: "Internal Server Error"
            message: "An unexpected error occurred"
            path: "/api/profiles"

  # ── SCHEMAS ───────────────────────────────────────────────
  schemas:
...
    UserSummaryDto:
      type: object
      description: Representación segura de un usuario (sin contraseña)
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          example: 1
        email:
          type: string
          format: email
          example: "cuidadora@ejemplo.com"
        fullName:
          type: string
          example: "María García"
        role:
          type: string
          enum: [USER, ADMIN]
          example: "USER"

    # ·· Profiles ··········································
    ProfileRequest:
      type: object
      description: Cuerpo para crear o actualizar un perfil de paciente
      properties:
        fullName:
          type: string
          example: "Juan Pérez López"
        email:
          type: string
          format: email
          example: "juan@ejemplo.com"
        birthDate:
          type: string
          format: date
          example: "1945-03-15"
        gender:
          type: string
          example: "Masculino"
        medicalConditions:
          type: string
          example: "Diabetes tipo 2, Hipertensión arterial"
        allergies:
          type: string
          example: "Penicilina"
        medications:
          type: string
          example: "Metformina 850mg, Enalapril 10mg"
        approxAddress:
          type: string
          example: "Calle Mayor 10, 2º B"
        city:
          type: string
          example: "Madrid"
        emergencyContactName:
          type: string
          example: "Ana Pérez"
        emergencyContactPhone:
          type: string
          example: "600123456"
        emergencyContactEmail:
          type: string
          format: email
          example: "ana@ejemplo.com"
        relationship:
          type: string
          example: "Hija"
        pinCode:
          type: string
          description: PIN de 4 dígitos para proteger el acceso público
          example: "1234"
...
    PrivateProfileDto:
      type: object
      description: Ficha completa del paciente (requiere autenticación JWT o PIN correcto)
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          example: 5
        fullName:
          type: string
          example: "Juan Pérez López"
        email:
          type: string
          format: email
          example: "juan@ejemplo.com"
        birthDate:
          type: string
          format: date
          example: "1945-03-15"
        gender:
          type: string
          example: "Masculino"
        medicalConditions:
          type: string
          example: "Diabetes tipo 2, Hipertensión arterial"
        allergies:
          type: string
          example: "Penicilina"
        medications:
          type: string
          example: "Metformina 850mg, Enalapril 10mg"
        approxAddress:
          type: string
          example: "Calle Mayor 10, 2º B"
        city:
          type: string
          example: "Madrid"
        emergencyContactName:
          type: string
          example: "Ana Pérez"
        emergencyContactPhone:
          type: string
          example: "600123456"
        emergencyContactEmail:
          type: string
          format: email
          example: "ana@ejemplo.com"
        relationship:
          type: string
          example: "Hija"
        qrToken:
          type: string
          readOnly: true
          description: Token único generado automáticamente al crear el perfil
          example: "a1b2c3d4e5f6g7h8i9j0"
        pinCode:
          type: string
          description: PIN de 4 dígitos para el acceso público mediante QR
          example: "1234"
        photoUrl:
          type: string
          format: uri
          readOnly: true
          example: "/images/a3f8b2c1.jpg"
        recentLogs:
          type: array
          readOnly: true
          description: Últimos registros de cuidado del paciente
          items:
            $ref: '#/components/schemas/CareLogDto'

    # ·· Care Logs ·········································
    CareLogRequest:
      type: object
      required: [logType]
      description: Cuerpo para crear un registro de cuidado (el campo `id` es ignorado)
      properties:
        logType:
          type: string
          description: Categoría del evento registrado
          enum: [MEDICACION, SINTOMAS, COMIDA, HIGIENE, OTRO]
          example: "MEDICACION"
        note:
          type: string
          description: Descripción libre del evento
          example: "Tomó ibuprofeno 400mg con el desayuno"
    CareLogDto:
      type: object
      description: Registro de cuidado persistido
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          example: 42
        profileId:
          type: integer
          format: int64
          readOnly: true
          example: 5
        logType:
          type: string
          example: "MEDICACION"
        note:
          type: string
          example: "Tomó ibuprofeno 400mg con el desayuno"
        createdAt:
          type: string
          format: date-time
          readOnly: true
          example: "2024-01-15T09:00:00"

    # ·· Reminders ·········································
    ReminderRequest:
      type: object
      required: [title, time]
      description: Cuerpo para crear un recordatorio
      properties:
        title:
          type: string
          example: "Pastilla tensión"
        time:
          type: string
          description: Hora del recordatorio en formato HH:mm:ss
          example: "08:30:00"
        daysOfWeek:
          type: string
          description: >
            Días de la semana separados por coma en los que se activa el recordatorio.
            Si se omite, se activa "TODOS" los días.
            Valores válidos: LUNES, MARTES, MIERCOLES, JUEVES, VIERNES, SABADO, DOMINGO, TODOS
          example: "LUNES,MIERCOLES,VIERNES"
    ReminderDto:
      type: object
      description: Recordatorio persistido
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          example: 7
        title:
          type: string
          example: "Pastilla tensión"
        time:
          type: string
          description: Hora en formato HH:mm:ss
          example: "08:30:00"
        isActive:
          type: boolean
          description: Indica si el recordatorio está habilitado
          example: true
        daysOfWeek:
          type: string
          description: Días activos separados por coma, o "TODOS"
          example: "LUNES,MIERCOLES,VIERNES"

    # ·· Public ············································
...
    UpdateRoleRequest:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [USER, ADMIN]
          description: Nuevo rol a asignar al usuario
          example: "ADMIN"

    # ·· Errors ············································
    ErrorResponse:
      type: object
      description: Formato estándar de error devuelto por ApiExceptionHandler
      properties:
        timestamp:
          type: string
          format: date-time
          description: Instante UTC en que ocurrió el error
          example: "2024-01-15T10:30:00Z"
        status:
          type: integer
          description: Código HTTP del error
          example: 400
        error:
          type: string
          description: Nombre corto del error HTTP
          example: "Bad Request"
        message:
          type: string
          description: Descripción legible del error
          example: "El email ya está registrado"
        path:
          type: string
          description: Ruta del endpoint que generó el error
          example: "/auth/register"
    ValidationErrorResponse:
      type: object
      description: Error 400 con detalle de qué campos fallaron la validación
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        status:
          type: integer
          example: 400
        error:
          type: string
          example: "Validation Error"
        message:
          type: string
          example: "Los datos enviados no son válidos"
        path:
          type: string
          example: "/auth/register"
        validationErrors:
          type: object
          description: Mapa campo → mensaje de error de validación
          additionalProperties:
            type: string
          example:
            email: "Debe ser un email valido"
            password: "La contraseña debe tener al menos 6 caracteres"
            fullName: "El nombre es obligatorio"